+++ packages/extension-polkagate/src/popup/accountDetailsFullScreen/components/index.ts
@@ -1,0 +1,11 @@
+// Copyright 2019-2023 @polkadot/extension-polkagate authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+export { default as AccountInformation } from './AccountInformation';
+export { default as AccountSetting } from './AccountSetting';
+export { default as ChangeAssets } from './ChangeAssets';
+export { default as CommonTasks } from './CommonTasks';
+export { default as DisplayBalance } from './DisplayBalance';
+export { default as ExternalLinks } from './ExternalLinks';
+export { default as LockedBalanceDisplay } from './LockedBalanceDisplay';
+export { default as TotalChart } from './TotalChart';
